Tests/LibWeb/Text/expected/css/animated-background-image-shared-resource-count.txt

timer active for shared animation: true
client count: 1
